refactor: move backend-specific Backend classes into per-backend `backend.py`
Move each backend `Backend` subclass out of the monolithic
`helion/_compiler/backend.py` into per-backend modules
`helion/_compiler/{triton,pallas,cute,metal}/backend.py` (TritonBackend +
TileIRBackend -> triton, PallasBackend -> pallas, CuteBackend -> cute,
MetalBackend -> metal), mirroring the existing per-backend separation.
Backend-exclusive auxiliaries move with their backend: SliceAddressing /
_slice_addressing / _TORCH_TO_JAX_DTYPE -> pallas; the cute-only MMA/flash
loop-detection helpers + _CUTE_DEFAULT_AUTOTUNE_BUDGET_SECONDS -> cute;
_triton_jit_supports_do_not_specialize -> triton. The base Backend class and
helpers shared across backends (e.g. _loop_contains_matmul, _active_loop_block_ids,
and the attention pattern-matching family reached from device_ir / tile_strategy)
stay in backend.py, which re-imports the subclasses at the bottom so
`from helion._compiler.backend import <Backend>` and backend_registry keep
resolving -- no behavior change. Repoints the moved-symbol imports in
pallas/tracing_ops.py and the affected tests.

stack-info: PR: https://github.com/pytorch/helion/pull/3090, branch: AmesingFlank/stack/75
